Remove bad-codeblock-syntax test
The test checks an ill-formed assumption. Lexer can't know if a bit of
code is well-or-ill formed. Even the old lexer didnt' error for most
of the code. rustc_lexer is capable of lexing everything.

The more straightforward logic is to just highlight everything. If we
want to guess if this is a valid rust or not, we'd better use a parser
for this.